chore: remove remove-use-strict
